Create index.html
This website gives access to 3 original pieces:

1. Horror Audio Track
This track made on Pro Tools is a short horror piece that employs vocals, percussion, and wind in order to create a spooky experience.

2. Simple Rap Beat
This beat uses all syntheiszed sound to create a simple beat that can be used or sampled to create anything. Created on Logic

3. "All I Want"
This is an original country song that uses recording of guitar, piano, and vocals all by me.